Broke core/{models.py,serializers.py} into their own python modules with serializers and models grouped.
