+++ backend/core/models/consumer.py
@@ -1,0 +1,13 @@
+from django.conf import settings
+from django.db import models
+from django.db.models import OneToOneField
+
+
+class Consumer(models.Model):
+    consumer_id = models.AutoField(primary_key=True)
+    user = OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
+    display_name = models.CharField(max_length=255)
+    streak = models.IntegerField(default=0)
+
+    class Meta:
+        db_table = "consumer"
