+++ backend/core/models/seller.py
@@ -1,0 +1,14 @@
+from django.conf import settings
+from django.db import models
+
+
+class Seller(models.Model):
+    seller_id = models.AutoField(primary_key=True)
+    user = models.OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
+    name = models.CharField(max_length=255)
+    location = models.CharField(max_length=255)
+    opening_hours = models.TextField(null=True, blank=True)
+    contact_stub = models.CharField(max_length=255, null=True, blank=True)
+
+    class Meta:
+        db_table = "seller"
