+++ backend/core/models/user.py
@@ -1,0 +1,6 @@
+from django.contrib.auth.models import AbstractUser
+from django.db import models
+
+
+class User(AbstractUser):
+    email = models.EmailField(max_length=255, unique=True)
