+++ backend/core/serializers/reports.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from core.models import IssueReport
+
+
+class IssueReportingSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = IssueReport
+        fields = "__all__"
